+++ libs/full/resiliency_distributed/examples/async_replicate_distributed.cpp
@@ -19,6 +19,7 @@
 #include <hpx/modules/testing.hpp>
 
 #include <cstddef>
+#include <cstdint>
 #include <iostream>
 #include <random>
 #include <vector>
@@ -26,7 +27,7 @@
 int universal_ans(std::vector<hpx::id_type> f_locales, std::size_t size)
 {
     // Pretending to do some useful work
-    std::size_t start = hpx::chrono::high_resolution_clock::now();
+    std::uint64_t start = hpx::chrono::high_resolution_clock::now();
 
     while ((hpx::chrono::high_resolution_clock::now() - start) < (size * 100))
     {
